제작창에서 돈 표시 구현

+++ Framework/Framework/WaveButton.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "WaveButton.h"
 #include "GameScene.h"
-WaveButton::WaveButton(const wchar_t* path, int colwidth, int colheight)
+WaveButton::WaveButton(const wchar_t* path, float colwidth, float colheight)
 :Button(path, colwidth, colheight)
 {
 }
