+++ Framework/Framework/WaveButton.h
@@ -4,6 +4,6 @@ class WaveButton :
 	public Button
 {
 public:
-	WaveButton(const wchar_t* path, int colwidth, int colheight);
+	WaveButton(const wchar_t* path, float colwidth, float colheight);
 	void OnClick();
 };
